version 1.3: - Data Screen: Columns don't expand to match label (But do expand to fill frame) - Save Mechanism: default name to "{name}-ACTIVE" - Fixed DJ CSV bug - Changed pages logic to two stages select-commit fully responsive: facets change as the variable selection change - Data Screen: moved data buttons and changed color, removed "select var." button - Tool tip for recoding changed - Changed memory file extension to .mms - Changed the hover color on the facet dim page to grey - Changed Radial Display to not extend behind the starting point - Added memory file to screen's title - Added dialogue on exiting the program

+++ lib/pages/input_page.py
@@ -217,5 +217,7 @@ class InputPage(ttk.Frame):
 
     def browse_file(self):
         filename = filedialog.askopenfilename()
-        self.entry_data_file.delete(0, ttk.END)
-        self.entry_data_file.insert(0, filename)
+        if filename:
+            self.entry_data_file.delete(0, ttk.END)
+            self.entry_data_file.insert(0, filename)
+        return filename
